Add seq2seq attack 1 record level
