+++ requirements.txt
@@ -1,10 +1,13 @@
 absl-py==0.9.0
 astor==0.8.1
 astroid==2.3.3
+autopep8==1.5.4
 cachetools==4.0.0
 certifi==2019.11.28
 chardet==3.0.4
+click==7.1.2
 cycler==0.10.0
+future-fstrings==1.2.0
 gast==0.2.2
 google-auth==1.11.0
 google-auth-oauthlib==0.4.1
@@ -13,25 +16,29 @@ grpcio==1.26.0
 h5py==2.10.0
 idna==2.8
 isort==4.3.21
+joblib==0.14.1
 Keras-Applications==1.0.8
 Keras-Preprocessing==1.1.0
 lazy-object-proxy==1.4.3
 Markdown==3.1.1
 matplotlib==2.0.2
 mccabe==0.6.1
+nltk==3.5
 numpy==1.18.1
 oauthlib==3.1.0
 opt-einsum==3.1.0
-pandas==0.25.3
+pandas==0.24.2
+progress==1.5
 protobuf==3.11.2
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
-pycodestyle==2.5.0
+pycodestyle==2.6.0
 pylint==2.4.4
 pyparsing==2.4.6
 python-dateutil==2.8.1
 pytz==2019.3
 PyYAML==5.1.1
+regex==2020.9.27
 requests==2.22.0
 requests-oauthlib==1.3.0
 rsa==4.0
@@ -40,9 +47,12 @@ scipy==1.4.1
 six==1.14.0
 sklearn==0.0
 tensorboard==2.1.1
-tensorflow-gpu==2.1.2
 tensorflow-estimator==2.1.0
+tensorflow-gpu==2.1.2
 termcolor==1.1.0
+tokenize-rt==3.2.0
+toml==0.10.1
+tqdm==4.50.2
 typed-ast==1.4.1
 urllib3==1.25.8
 Werkzeug==0.16.0
