+++ tutorials/README.md
@@ -22,7 +22,7 @@ The `attackobj` initializes the meminf class and the attack configuration. Follo
 Note : The code explicitly sets the means and standard deviations for normalizing the images, according to the CIFAR-100 distribution.
 1. Whitebox attack - Exploit the final layer gradients, final layer outputs, loss values and label values (DEFAULT)
 ```
-attackobj = ml_privacy_meter.attack.meminf.initialize(
+attackobj = ml_privacy_meter.attack.meminf.meminf(
     target_train_model=cmodelA,
     target_attack_model=cmodelA,
     train_datahandler=datahandlerA,
@@ -33,7 +33,7 @@ attackobj = ml_privacy_meter.attack.meminf.initialize(
 ```
 2. Whitebox attack - Exploit final two model layer outputs, loss values and label values
 ```
-attackobj = ml_privacy_meter.attack.meminf.initialize(
+attackobj = ml_privacy_meter.attack.meminf.meminf(
     target_train_model=cmodelA,
     target_attack_model=cmodelA,
     train_datahandler=datahandlerA,
@@ -43,7 +43,7 @@ attackobj = ml_privacy_meter.attack.meminf.initialize(
 ```
 2. Blackbox attack - Exploit final layer output and label values
 ```
-attackobj = ml_privacy_meter.attack.meminf.initialize(
+attackobj = ml_privacy_meter.attack.meminf.meminf(
     target_train_model=cmodelA,
     target_attack_model=cmodelA,
     train_datahandler=datahandlerA,
